Refactor: Separate MCP client concerns according to design document
- Split McpClient into generic MCP operations only
- Create IdobataMcpService for business logic and OpenAI integration
- Replace try-catch with neverthrow early return patterns
- Update ProcessChatMessageUsecase to use new architecture
- Move OpenAI integration from McpClient to IdobataMcpService
- Maintain backward compatibility with existing interfaces
