+++ policy-edit/backend/src/usecases/ConnectMcpServerUsecase.ts
@@ -38,36 +38,28 @@ export class ConnectMcpServerUsecase {
   private async cleanupExistingConnection(): Promise<
     Result<void, McpClientError>
   > {
-    try {
-      if (this.mcpClientRef.current) {
-        await this.mcpClientRef.current.cleanup();
+    if (this.mcpClientRef.current) {
+      const cleanupResult = await this.mcpClientRef.current.cleanup();
+      if (cleanupResult.isErr()) {
+        return err(cleanupResult.error);
       }
-      return ok(undefined);
-    } catch (error) {
-      return err(
-        new McpClientError(
-          `Failed to cleanup existing connection: ${error instanceof Error ? error.message : "Unknown error"}`
-        )
-      );
     }
+    return ok(undefined);
   }
 
   private async establishConnection(
     serverPath: string
   ): Promise<Result<void, McpClientError>> {
-    try {
-      this.mcpClientRef.current = new McpClient();
-      await this.mcpClientRef.current.connectToServer(serverPath);
-      logger.info(`MCP client connected to server at ${serverPath}`);
-      return ok(undefined);
-    } catch (error) {
-      logger.error("Failed to initialize MCP client:", error);
+    this.mcpClientRef.current = new McpClient();
+    const connectResult = await this.mcpClientRef.current.connectToServer(serverPath);
+    
+    if (connectResult.isErr()) {
+      logger.error("Failed to initialize MCP client:", connectResult.error);
       this.mcpClientRef.current = null;
-      return err(
-        new McpClientError(
-          `Failed to connect to MCP server: ${error instanceof Error ? error.message : "Unknown error"}`
-        )
-      );
+      return err(connectResult.error);
     }
+    
+    logger.info(`MCP client connected to server at ${serverPath}`);
+    return ok(undefined);
   }
 }
